requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask~=3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
Werkzeug>=3.1
requests~=2.32.3
selenium~=4.29.0
pillow~=11.1.0
beautifulsoup4~=4.13.3